Overhaul Keyboard Menu Navigation.
The old navigation worked similarly to how tab+shift tab works on most websites, except it also handled wasd and arrow keys, it would take a list of all available actions and let you step through the list.

This actually worked completely fine for most pages. The only issue was that if there was a button to your left and right, and a button below you, your inputs wouldn't allow you to choose to go any direction, instead, you would follow a specific route throughout the page.

Instead of using an array to hold each action, the new navigation uses a tree-like structure (objects that have data about the HTML element they represent and a "children" property with more objects representing elements).

This gives us the ability to tell who our siblings are as well as our parents, and lets us step in any direction.

I will add more information on the comment under this commit.