update car to have same physics while backwards
the physics are now the same in both directions. BUT its missing the ability to be backwards, get into a drift, then spin and go forwards.

All I really added was the ability to drift backwards as well as forwards.

Since the car physics don't use velocity vectors:  If you begin a drift with a negative speed, it will always try to recorrect you to face your original forward direction.